rearrange api actions into controllers. handle auth errors on client
